Add JWT claim

+++ resources/user.py
@@ -5,7 +5,7 @@ from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from passlib.hash import pbkdf2_sha256
 
-from flask_jwt_extended import create_access_token
+from flask_jwt_extended import create_access_token, get_jwt
 
 from models import UserModel
 
@@ -75,6 +75,11 @@ class User(MethodView):
             abort(404, message="users not found")
 
     def delete(self, user_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required")
+        
+        
         user = UserModel.query.get_or_404(user_id)
         
         db.session.delete(user)
